requirements.txt
aria2p==0.10.4
google-api-python-client==2.7.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
psutil==5.8.0
python-magic==0.4.24
python-telegram-bot==13.5
